tests: Move stream permission tests to test_channel_permissions.
